errors: make error messages more helpful
Some API Developers have reported AWeberOAuthDataMissing
exceptions when using the demo.php script.

This error message is not helpful as the typical cause for this
exception is an invalid consumer key or secret.

The client library has been refactored to always raise an
AWeberAPIException when a 40x/50x http status code is returned.

This exception will clearly indicate the error for troubleshooting.