Fix handling of reserved chars when generating oauth signatures
Clearly define how reserved characters are to be encoded in calls
to OAuthApplication->makeRequest. Treat reserved characters as
though they are URL encoded when they are passed in as query params
to the URL. Treat reserved characters as though they are not URL
encoded when passed in as an entry in a data array. Generate unit
tests to prove that the signature is being generated correctly.